modularizacion de componentes
