+++ src/App.tsx
@@ -6,9 +6,9 @@ import "@fontsource/outfit/700.css";  // Bold
 import "@fontsource/inter/400.css";   // Para textos regulares
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Navbar from "./components/Navbar"
-import Home from './components/Home';
-import Services from './components/Services';
-import Contact from "./components/Contact";
+import Home from './pages/Home';
+import Services from './pages/Services';
+import Contact from "./pages/Contact";
 
 function App() {
   return (
