+++ src/types/index.ts
@@ -1,0 +1,7 @@
+export interface DetailedService {
+    title: string;
+    description: string;
+    icon: string;
+    features: string[];
+    benefits: string[];
+}
