+++ src/utils/data.ts
@@ -1,0 +1,107 @@
+import { DetailedService } from "../types/index"
+
+export const detailedServices: DetailedService[] = [
+    {
+        title: "Consultoría Tecnológica",
+        description: "Asesoramiento estratégico para optimizar y transformar tu infraestructura tecnológica.",
+        icon: "/icons/contec.svg",
+        features: [
+            "Evaluación de infraestructura actual",
+            "Planificación estratégica tecnológica",
+            "Optimización de recursos IT",
+            "Roadmap de transformación digital"
+        ],
+        benefits: [
+            "Reducción de costos operativos",
+            "Mejora en la eficiencia de procesos",
+            "Ventaja competitiva tecnológica",
+            "Escalabilidad planificada"
+        ]
+    },
+    {
+        title: "Automatización Inteligente",
+        description: "Optimización de procesos mediante tecnologías avanzadas de automatización e IA.",
+        icon: "/icons/autint.svg",
+        features: [
+            "Automatización de procesos repetitivos",
+            "Implementación de IA y Machine Learning",
+            "Workflows personalizados",
+            "Integración con sistemas existentes"
+        ],
+        benefits: [
+            "Reducción de errores humanos",
+            "Mayor eficiencia operativa",
+            "Optimización de recursos",
+            "Análisis predictivo"
+        ]
+    },
+    {
+        title: "Desarrollo a Medida",
+        description: "Creación de soluciones tecnológicas adaptadas perfectamente a tus necesidades específicas.",
+        icon: "icons/desmed.svg",
+        features: [
+            "Desarrollo de software personalizado",
+            "Aplicaciones web y móviles",
+            "Integración de sistemas",
+            "Arquitectura escalable"
+        ],
+        benefits: [
+            "Soluciones 100% adaptadas",
+            "Mayor control y flexibilidad",
+            "Ventaja competitiva única",
+            "Evolución continua del software"
+        ]
+    },
+    {
+        title: "Administración de Infraestructura",
+        description: "Gestión integral y optimización de infraestructura tecnológica para máximo rendimiento.",
+        icon: "icons/admininfra.svg"
+        ,
+        features: [
+            "Gestión de servidores y redes",
+            "Seguridad y respaldo de datos",
+            "Optimización de recursos",
+            "Mantenimiento preventivo"
+        ],
+        benefits: [
+            "Alta disponibilidad",
+            "Seguridad reforzada",
+            "Rendimiento optimizado",
+            "Reducción de incidentes"
+        ]
+    },
+    {
+        title: "Monitoreo y Análisis",
+        description: "Supervisión continua y análisis en tiempo real de sistemas críticos.",
+        icon: "icons/monanalisis.svg",
+        features: [
+            "Monitoreo 24/7",
+            "Alertas en tiempo real",
+            "Análisis de tendencias",
+            "Reportes personalizados"
+        ],
+        benefits: [
+            "Prevención de incidentes",
+            "Tiempo de respuesta mejorado",
+            "Decisiones basadas en datos",
+            "Optimización continua"
+        ]
+    },
+    {
+        title: "ETL y Business Intelligence",
+        description: "Transformamos datos en conocimiento mediante procesos ETL y análisis avanzado.",
+        icon: "icons/etlbi.svg",
+        features: [
+            "Integración de fuentes de datos",
+            "Transformación y limpieza",
+            "Visualización avanzada",
+            "Dashboards interactivos"
+        ],
+        benefits: [
+            "Decisiones informadas",
+            "Insights en tiempo real",
+            "Ventaja competitiva analítica",
+            "ROI mejorado"
+        ]
+    }
+];
